scenario-doc/requirements.txt
PyYAML  # MIT
sphinx>=2.0.0 # BSD
